+++ src/main/java/stackpot/stackpot/pot/service/MyPotService.java
@@ -1,51 +1,21 @@
 package stackpot.stackpot.pot.service;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
+
 import stackpot.stackpot.badge.dto.CompletedPotBadgeResponseDto;
-import stackpot.stackpot.task.dto.*;
-import stackpot.stackpot.task.entity.enums.TaskboardStatus;
 import stackpot.stackpot.pot.dto.CompletedPotDetailResponseDto;
 import stackpot.stackpot.pot.dto.OngoingPotResponseDto;
-import stackpot.stackpot.todo.dto.MyPotTodoRequestDTO;
-import stackpot.stackpot.todo.dto.MyPotTodoResponseDTO;
-import stackpot.stackpot.todo.dto.MyPotTodoUpdateRequestDTO;
 
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Map;
 
 public interface MyPotService {
 
     // 사용자의 진행 중인 팟 조회
     List<OngoingPotResponseDto> getMyPots();
-    // 사용자의 특정 팟에서의 생성
-    List<MyPotTodoResponseDTO> postTodo(Long potId, MyPotTodoRequestDTO requestDTO);
-
-    Page<MyPotTodoResponseDTO> getTodo(Long potId, PageRequest pageRequest);
-
-    List<MyPotTodoResponseDTO> updateTodos(Long potId, List<MyPotTodoUpdateRequestDTO> requestList);
-
-    List<MyPotTodoResponseDTO> completeTodo(Long potId, Long todoId);
-
-    MyPotTaskResponseDto creatTask(Long potId, MyPotTaskRequestDto.create request);
-
-    public Map<TaskboardStatus, List<MyPotTaskPreViewResponseDto>> preViewTask(Long potId);
-    MyPotTaskResponseDto viewDetailTask(Long potId, Long taskboardId);
-    void deleteTaskboard(Long potId, Long taskboardId);
     CompletedPotDetailResponseDto getCompletedPotDetail(Long potId);
     List<CompletedPotBadgeResponseDto> getCompletedPotsWithBadges();
     List<CompletedPotBadgeResponseDto> getUserCompletedPotsWithBadges(Long userId);
     List<OngoingPotResponseDto> getMyOngoingPots();
-
-    MyPotTaskStatusResponseDto updateTaskStatus(Long potId, Long taskId, TaskboardStatus status);
-    List<MyPotTaskPreViewResponseDto> getTasksFromDate(Long potId, LocalDate date);
-
-    List<MonthlyTaskDto> getMonthlyTasks(Long potId, int year, int month);
-
     boolean isOwner(Long potId);
-
-    MyPotTaskResponseDto modifyTask(Long potId, Long taskboardId, MyPotTaskRequestDto.create request);
 
 
 }
