+++ src/main/java/stackpot/stackpot/pot/service/MyPotServiceImpl.java
@@ -3,9 +3,6 @@ package stackpot.stackpot.pot.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -14,36 +11,19 @@ import stackpot.stackpot.apiPayload.exception.handler.MemberHandler;
 import stackpot.stackpot.apiPayload.exception.handler.PotHandler;
 import stackpot.stackpot.badge.dto.BadgeDto;
 import stackpot.stackpot.badge.dto.CompletedPotBadgeResponseDto;
+import stackpot.stackpot.common.util.RoleNameMapper;
 import stackpot.stackpot.pot.converter.MyPotConverter;
-import stackpot.stackpot.pot.converter.PotConverter;
 import stackpot.stackpot.pot.converter.PotDetailConverter;
-import stackpot.stackpot.task.converter.TaskboardConverter;
 import stackpot.stackpot.pot.dto.CompletedPotDetailResponseDto;
 import stackpot.stackpot.pot.dto.OngoingPotResponseDto;
 import stackpot.stackpot.pot.entity.Pot;
-import stackpot.stackpot.task.dto.*;
-import stackpot.stackpot.task.entity.Taskboard;
-import stackpot.stackpot.todo.dto.MyPotTodoRequestDTO;
-import stackpot.stackpot.todo.dto.MyPotTodoResponseDTO;
-import stackpot.stackpot.todo.dto.MyPotTodoUpdateRequestDTO;
 import stackpot.stackpot.user.entity.User;
 import stackpot.stackpot.user.entity.enums.Role;
-import stackpot.stackpot.task.entity.enums.TaskboardStatus;
-import stackpot.stackpot.todo.entity.enums.TodoStatus;
 import stackpot.stackpot.pot.entity.mapping.PotMember;
-import stackpot.stackpot.task.entity.mapping.Task;
-import stackpot.stackpot.todo.entity.mapping.UserTodo;
 import stackpot.stackpot.badge.repository.PotMemberBadgeRepository;
 import stackpot.stackpot.pot.repository.PotMemberRepository;
-import stackpot.stackpot.pot.repository.MyPotRepository;
 import stackpot.stackpot.pot.repository.PotRepository;
-import stackpot.stackpot.task.repository.TaskRepository;
-import stackpot.stackpot.task.repository.TaskboardRepository;
 import stackpot.stackpot.user.repository.UserRepository;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -54,13 +34,8 @@ import java.util.stream.Collectors;
 public class MyPotServiceImpl implements MyPotService {
 
     private final PotRepository potRepository;
-    private final MyPotRepository myPotRepository;
     private final UserRepository userRepository;
-    private final PotConverter potConverter;
-    private final TaskboardConverter taskboardConverter;
-    private final TaskboardRepository taskboardRepository;
     private final PotMemberRepository potMemberRepository;
-    private final TaskRepository taskRepository;
     private final PotDetailConverter potDetailConverter;
     private final MyPotConverter myPotConverter;
     private final PotMemberBadgeRepository potMemberBadgeRepository;
@@ -100,472 +75,6 @@ public class MyPotServiceImpl implements MyPotService {
     }
 
 
-    @Override
-    public List<MyPotTodoResponseDTO> postTodo(Long potId, MyPotTodoRequestDTO requestDTO) {
-        // 현재 인증된 사용자 가져오기
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        // 사용자 정보 조회
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-        // 팟 조회
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new PotHandler(ErrorStatus.POT_NOT_FOUND));
-
-        // 소유자 또는 멤버 권한 확인
-        boolean isOwner = pot.getUser().equals(user);
-        boolean isMember = potMemberRepository.existsByPotAndUser(pot, user); // 팟의 멤버 여부 확인
-
-        if (!isOwner && !isMember) {
-            throw new PotHandler(ErrorStatus.POT_FORBIDDEN); // 권한 없음
-        }
-
-        // To-Do 생성
-        UserTodo userTodo = UserTodo.builder()
-                .pot(pot)
-                .user(user)
-                .content(requestDTO.getContent())
-                .status(requestDTO.getStatus() != null ? requestDTO.getStatus() : TodoStatus.NOT_STARTED)
-                .build();
-
-        myPotRepository.save(userTodo);
-
-        // 특정 팟의 모든 To-Do 조회 (업데이트된 리스트)
-        List<UserTodo> potTodos = myPotRepository.findByPot_PotId(potId);
-
-        //  투두를 User 기준으로 그룹핑
-        Map<User, List<UserTodo>> groupedByUser = potTodos.stream()
-                .collect(Collectors.groupingBy(UserTodo::getUser));
-
-        // 사용자별로 그룹화하여 반환
-        return groupedByUser.entrySet().stream()
-                .map(entry -> {
-                    User currentUser = entry.getKey();
-                    String roleName = getUserRoleInPot(currentUser, pot);
-                    List<UserTodo> userTodos = groupedByUser.getOrDefault(currentUser, List.of());
-
-                    return MyPotTodoResponseDTO.builder()
-                            .userNickname(currentUser.getNickname() + getVegetableNameByRole(roleName))
-                            .userId(currentUser.getId())
-                            .todoCount(userTodos.size())
-                            .todos(userTodos.stream()
-                                    .map(todo -> MyPotTodoResponseDTO.TodoDetailDTO.builder()
-                                            .todoId(todo.getTodoId())
-                                            .content(todo.getContent())
-                                            .status(todo.getStatus())
-                                            .build())
-                                    .collect(Collectors.toList()))
-                            .build();
-                })
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public Page<MyPotTodoResponseDTO> getTodo(Long potId, PageRequest pageRequest) {
-        // 현재 인증된 사용자 가져오기
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-        // 팟 조회
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new PotHandler(ErrorStatus.POT_NOT_FOUND));
-
-        // 소유자 또는 멤버 권한 확인
-        boolean isOwner = pot.getUser().equals(user);
-        boolean isMember = potMemberRepository.existsByPotAndUser(pot, user);
-
-        if (!isOwner && !isMember) {
-            throw new PotHandler(ErrorStatus.POT_FORBIDDEN);
-        }
-
-        // 팟의 모든 멤버 조회 (소유자 포함) 후, User 기준으로 페이징
-        List<User> allPotMembers = potMemberRepository.findByPotId(pot.getPotId())
-                .stream()
-                .map(PotMember::getUser)
-                .collect(Collectors.toList());
-
-        allPotMembers.sort((u1, u2) -> u1.equals(user) ? -1 : (u2.equals(user) ? 1 : 0));
-
-        // User 기준으로 페이징 적용
-        int totalUsers = allPotMembers.size();
-        int startIndex = (int) pageRequest.getOffset();
-        int endIndex = Math.min(startIndex + pageRequest.getPageSize(), totalUsers);
-
-        if (startIndex >= totalUsers) {
-            return new PageImpl<>(List.of(), pageRequest, totalUsers);
-        }
-
-        List<User> pagedUsers = allPotMembers.subList(startIndex, endIndex);
-
-        // 선택된 User들의 투두 조회
-        List<UserTodo> todos = myPotRepository.findByPotAndUsers(pot, pagedUsers);
-
-        // createdAt이 전날 새벽 3시 이후인 것만 필터링
-        LocalDateTime now = LocalDateTime.now();
-        LocalDateTime todayAt3AM = LocalDateTime.of(now.toLocalDate(), LocalTime.of(3, 00));
-        LocalDateTime yesterdayAt3AM = todayAt3AM.minusDays(1);
-
-        List<UserTodo> filteredTodos = todos.stream()
-                .filter(todo -> todo.getCreatedAt().isAfter(yesterdayAt3AM))
-                .collect(Collectors.toList());
-
-        // 투두를 User 기준으로 그룹핑
-        Map<User, List<UserTodo>> groupedByUser = filteredTodos.stream()
-                .collect(Collectors.groupingBy(UserTodo::getUser));
-
-        // DTO 변환 (Converter 사용)
-        List<MyPotTodoResponseDTO> responseList = pagedUsers.stream()
-                .map(member -> myPotConverter.toDto(
-                        member,
-                        pot,
-                        groupedByUser.getOrDefault(member, List.of()),
-                        user))
-                .collect(Collectors.toList());
-
-        // Page 객체로 변환하여 반환
-        return new PageImpl<>(responseList, pageRequest, totalUsers);
-    }
-
-
-    @Override
-    @Transactional
-    public List<MyPotTodoResponseDTO> updateTodos(Long potId, List<MyPotTodoUpdateRequestDTO> requestList) {
-        // 현재 인증된 사용자 가져오기
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        // 사용자 정보 조회
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-        // 팟 조회
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new PotHandler(ErrorStatus.POT_NOT_FOUND));
-
-        List<UserTodo> existingTodos = myPotRepository.findByPot_PotIdAndUser(potId, user);
-
-        // 요청된 todoId 리스트
-        Set<Long> requestedTodoIds = requestList.stream()
-                .map(MyPotTodoUpdateRequestDTO::getTodoId)
-                .filter(Objects::nonNull) // 새로 추가된 항목(null) 제외
-                .collect(Collectors.toSet());
-
-        Map<Long, UserTodo> existingTodoMap = existingTodos.stream()
-                .collect(Collectors.toMap(UserTodo::getTodoId, todo -> todo));
-
-        List<UserTodo> updatedOrNewTodos = new ArrayList<>();
-
-        for (MyPotTodoUpdateRequestDTO request : requestList) {
-            if (request.getTodoId() != null && existingTodoMap.containsKey(request.getTodoId())) {
-                // 기존 투두 수정 (status 변경 없이 유지)
-                UserTodo existingTodo = existingTodoMap.get(request.getTodoId());
-
-                if (!existingTodo.getUser().equals(user)) {
-                    throw new PotHandler(ErrorStatus.USER_TODO_UNAUTHORIZED);
-                }
-
-                existingTodo.setContent(request.getContent());
-                // 기존 상태 유지 (status 변경 없음)
-                updatedOrNewTodos.add(existingTodo);
-            } else {
-                // 새로운 투두 생성
-                UserTodo newTodo = UserTodo.builder()
-                        .user(user)
-                        .pot(pot)
-                        .content(request.getContent())
-                        .status(request.getStatus() != null ? request.getStatus() : TodoStatus.NOT_STARTED) // 기본값 설정
-                        .build();
-                updatedOrNewTodos.add(newTodo);
-            }
-        }
-
-        // 삭제할 기존 투두 (삭제 요청이 없는 것만 남김)
-        List<UserTodo> todosToDelete = existingTodos.stream()
-                .filter(todo -> !requestedTodoIds.contains(todo.getTodoId()) && todo.getUser().equals(user)) // 본인만 삭제 가능
-                .collect(Collectors.toList());
-
-        myPotRepository.saveAll(updatedOrNewTodos);
-        myPotRepository.deleteAll(todosToDelete);
-
-        Map<User, List<UserTodo>> groupedByUser = updatedOrNewTodos.stream()
-                .collect(Collectors.groupingBy(UserTodo::getUser));
-
-        return groupedByUser.entrySet().stream()
-                .map(entry -> {
-                    // 해당 유저의 pot에서 potMember 역할 찾기
-                    String roleName = getUserRoleInPot(entry.getKey(), pot);
-                    List<UserTodo> userTodos = groupedByUser.getOrDefault(entry.getKey(), List.of());
-
-                    return MyPotTodoResponseDTO.builder()
-                            .userNickname(entry.getKey().getNickname() + getVegetableNameByRole(roleName))
-                            .userId(entry.getKey().getId())
-                            .todoCount(userTodos.size())
-                            .todos(userTodos.stream()
-                                    .map(todo -> MyPotTodoResponseDTO.TodoDetailDTO.builder()
-                                            .todoId(todo.getTodoId())
-                                            .content(todo.getContent())
-                                            .status(todo.getStatus()) // 기존 상태 유지
-                                            .build())
-                                    .collect(Collectors.toList()))
-                            .build();
-                })
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    @Override
-    public List<MyPotTodoResponseDTO> completeTodo(Long potId, Long todoId) {
-        // 현재 로그인한 사용자 확인
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        // 사용자 정보 조회
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-        // 팟 조회
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new PotHandler(ErrorStatus.POT_NOT_FOUND));
-
-        // 해당 투두가 존재하는지 확인 및 소유자 검증
-        UserTodo userTodo = myPotRepository.findByTodoIdAndPot_PotId(todoId, potId)
-                .orElseThrow(() -> new PotHandler(ErrorStatus.USER_TODO_NOT_FOUND));
-
-        if (!userTodo.getUser().equals(user)) {
-            throw new PotHandler(ErrorStatus.USER_TODO_UNAUTHORIZED);
-        }
-
-        // 기존 상태 확인 후 토글
-        if (userTodo.getStatus() == TodoStatus.COMPLETED) {
-            userTodo.setStatus(TodoStatus.NOT_STARTED);
-        } else {
-            userTodo.setStatus(TodoStatus.COMPLETED);
-        }
-
-        myPotRepository.save(userTodo);
-
-        // 특정 팟의 모든 To-Do 조회 후 반환
-        List<UserTodo> potTodos = myPotRepository.findByPot_PotId(potId);
-
-        Map<User, List<UserTodo>> groupedByUser = potTodos.stream()
-                .collect(Collectors.groupingBy(UserTodo::getUser));
-
-        return groupedByUser.entrySet().stream()
-                .map(entry -> {
-                    String roleName = getUserRoleInPot(entry.getKey(), pot);
-                    String userNicknameWithRole = entry.getKey().getNickname() + getVegetableNameByRole(roleName);
-                    List<UserTodo> userTodos = groupedByUser.getOrDefault(entry.getKey(), List.of());
-
-                    return MyPotTodoResponseDTO.builder()
-                            .userNickname(userNicknameWithRole)
-                            .userId(entry.getKey().getId())
-                            .todoCount(userTodos.size())
-                            .todos(entry.getValue().stream()
-                                    .map(todo -> MyPotTodoResponseDTO.TodoDetailDTO.builder()
-                                            .todoId(todo.getTodoId())
-                                            .content(todo.getContent())
-                                            .status(todo.getStatus())
-                                            .build())
-                                    .collect(Collectors.toList()))
-                            .build();
-                })
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public MyPotTaskResponseDto creatTask(Long potId, MyPotTaskRequestDto.create request) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new IllegalArgumentException("Pot not found with id: " + potId));
-
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(()-> new IllegalArgumentException("User not found with email: " + email));
-
-        Taskboard taskboard = taskboardConverter.toTaskboard(pot, request);
-        taskboard.setUser(user);
-        taskboardRepository.save(taskboard);
-
-        //List<Long> requestedParticipantIds = request.getParticipants();
-        List<Long> requestedParticipantIds = request.getParticipants() != null ? request.getParticipants() : List.of();
-
-
-        List<PotMember> validParticipants = potMemberRepository.findByPotId(potId);
-
-        List<PotMember> participants = validParticipants.stream()
-                .filter(potMember -> requestedParticipantIds.contains(potMember.getPotMemberId()))
-                .collect(Collectors.toList());
-
-        log.info("유효한 참가자 목록: {}", participants);
-
-        /*if (participants.isEmpty()) {
-            throw new IllegalArgumentException("유효한 참가자를 찾을 수 없습니다. 요청된 ID를 확인해주세요.");
-        }*/
-        createAndSaveTasks(taskboard, participants);
-
-        List<MyPotTaskResponseDto.Participant> participantDtos = taskboardConverter.toParticipantDtoList(participants);
-        MyPotTaskResponseDto response = taskboardConverter.toDTO(taskboard, participants);
-        response.setParticipants(participantDtos);
-
-        return response;
-    }
-
-    @Override
-    public Map<TaskboardStatus, List<MyPotTaskPreViewResponseDto>> preViewTask(Long potId) {
-
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(()->new IllegalArgumentException("Pot not found with id: " + potId));
-
-        List<Taskboard> taskboards = taskboardRepository.findByPot(pot);
-
-        List<MyPotTaskPreViewResponseDto> taskboardDtos = taskboards.stream()
-                .map(taskboard -> {
-                    List<Task> tasks = taskRepository.findByTaskboard(taskboard); // Task 조회
-                    List<PotMember> participants = tasks.stream()
-                            .map(Task::getPotMember) // Task에서 PotMember 추출
-                            .distinct()
-                            .collect(Collectors.toList());
-
-                    return taskboardConverter.toDto(taskboard, participants);
-                })
-                .collect(Collectors.toList());
-
-
-        return taskboardDtos.stream()
-                .collect(Collectors.groupingBy(MyPotTaskPreViewResponseDto::getStatus));
-    }
-
-
-    @Override
-    public MyPotTaskResponseDto viewDetailTask(Long potId, Long taskboardId) {
-
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(()-> new IllegalArgumentException("pot을 찾을 수 없습니다."));
-
-        Taskboard taskboard = taskboardRepository.findByPotAndTaskboardId(pot, taskboardId);
-        if (taskboard == null) {
-            throw new IllegalArgumentException("taskboard를 찾을 수 없습니다.");
-        }
-
-        List<Task> tasks = taskRepository.findByTaskboard(taskboard);
-
-        List<PotMember> participants = tasks.stream()
-                .map(Task::getPotMember) // Task에서 PotMember 추출
-                .distinct()
-                .collect(Collectors.toList());
-
-        MyPotTaskResponseDto response = taskboardConverter.toDTO(taskboard,participants);
-
-        return response;
-    }
-
-    @Override
-    @Transactional
-    public MyPotTaskResponseDto modifyTask(Long potId, Long taskboardId, MyPotTaskRequestDto.create request) {
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new IllegalArgumentException("pot을 찾을 수 없습니다."));
-
-        Taskboard taskboard = taskboardRepository.findByPotAndTaskboardId(pot, taskboardId);
-        if (taskboard == null) {
-            throw new IllegalArgumentException("taskboard를 찾을 수 없습니다.");
-        }
-
-        updateUserData(taskboard, request);
-
-        // ✅ participants가 null이면 빈 리스트로 처리
-        List<Long> requestedParticipantIds = request.getParticipants() != null ? request.getParticipants() : List.of();
-        List<PotMember> participants = potMemberRepository.findAllById(requestedParticipantIds);
-
-        log.info("참가자 {}", participants);
-
-        // ✅ 기존 테스크 삭제 (참여자가 없어도 삭제되어야 함)
-        taskRepository.deleteByTaskboard(taskboard);
-
-        // ✅ 새로운 참여자가 있을 경우에만 Task 생성
-        if (!participants.isEmpty()) {
-            createAndSaveTasks(taskboard, participants);
-        } else {
-            log.info("참여자가 없어 기존 테스크 삭제 후 새로운 Task는 생성되지 않습니다.");
-        }
-
-        List<MyPotTaskResponseDto.Participant> participantDtos = taskboardConverter.toParticipantDtoList(participants);
-        MyPotTaskResponseDto response = taskboardConverter.toDTO(taskboard, participants);
-        response.setParticipants(participantDtos);
-
-        return response;
-    }
-
-    @Transactional
-    @Override
-    public void deleteTaskboard(Long potId, Long taskboardId) {
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new IllegalArgumentException("pot을 찾을 수 없습니다."));
-
-        Taskboard taskboard = taskboardRepository.findByPotAndTaskboardId(pot, taskboardId);
-        if (taskboard == null) {
-            throw new IllegalArgumentException("taskboard를 찾을 수 없습니다.");
-        }
-
-        // Taskboard에 연결된 Task 삭제
-        List<Task> tasks = taskRepository.findByTaskboard(taskboard);
-        taskRepository.deleteAll(tasks);
-
-        // Taskboard 삭제
-        taskboardRepository.delete(taskboard);
-    }
-
-    // 역할에 따른 채소명을 반환하는 메서드
-    private String getVegetableNameByRole(String role) {
-        Map<String, String> roleToVegetableMap = Map.of(
-                "BACKEND", " 양파",
-                "FRONTEND", " 버섯",
-                "DESIGN", " 브로콜리",
-                "PLANNING", " 당근",
-                "UNKNOWN",""
-        );
-        return roleToVegetableMap.getOrDefault(role, "알 수 없음");
-    }
-
-
-    private String getUserRoleInPot(User user, Pot pot) {
-        return pot.getPotMembers().stream()
-                .filter(member -> member.getUser().equals(user))
-                .map(member -> member.getRoleName().name())  // ENUM -> String 변환
-                .findFirst()
-                .orElse("UNKNOWN");  // 기본값 설정
-    }
-
-    private void updateUserData(Taskboard taskboard, MyPotTaskRequestDto.create request) {
-        if(request.getTitle() !=null){
-            taskboard.setTitle(request.getTitle());
-        }
-        if(request.getDescription()!=null){
-            taskboard.setDescription(request.getDescription());
-        }
-        if(request.getDeadline()!=null){
-            taskboard.setDeadLine(request.getDeadline());
-        }
-        if(request.getTaskboardStatus()!=null){
-            taskboard.setStatus(request.getTaskboardStatus());
-        }
-    }
-
-    private List<Task> createAndSaveTasks(Taskboard taskboard, List<PotMember> participants) {
-        List<Task> tasks = participants.stream()
-                .map(participant -> Task.builder()
-                        .taskboard(taskboard)
-                        .potMember(participant)
-                        .build())
-                .collect(Collectors.toList());
-
-        return taskRepository.saveAll(tasks);
-    }
-
     @Transactional
     @Override
     public CompletedPotDetailResponseDto getCompletedPotDetail(Long potId) {
@@ -596,8 +105,8 @@ public class MyPotServiceImpl implements MyPotService {
 
         // Pot 멤버의 Role 조회 후 변환
         userPotRole = potMemberRepository.findRoleByUserId(pot.getPotId(), user.getId())
-                .map(role -> getKoreanRoleName(role.name())) //  Optional<Role>을 String으로 변환 후 한글 적용
-                .orElse(getKoreanRoleName(pot.getUser().getRole().name())); // 기본값: Pot 생성자의 Role
+                .map(role -> RoleNameMapper.getKoreanRoleName(role.name()))
+                .orElse(RoleNameMapper.getKoreanRoleName(pot.getUser().getRole().name()));
 
 
         // DTO 반환
@@ -632,7 +141,7 @@ public class MyPotServiceImpl implements MyPotService {
 
                     //  "프론트엔드(2), 백엔드(1)" 형식으로 변환
                     String formattedMembers = roleCountsMap.entrySet().stream()
-                            .map(entry -> getKoreanRoleName(entry.getKey()) + "(" + entry.getValue() + ")")
+                            .map(entry -> RoleNameMapper.getKoreanRoleName(entry.getKey()) + "(" + entry.getValue() + ")")
                             .collect(Collectors.joining(", "));
 
                     Role userPotRole = potMemberRepository.findRoleByUserId(pot.getPotId(), user.getId())
@@ -673,7 +182,7 @@ public class MyPotServiceImpl implements MyPotService {
 
                     //  "프론트엔드(2), 백엔드(1)" 형식으로 변환
                     String formattedMembers = roleCountsMap.entrySet().stream()
-                            .map(entry -> getKoreanRoleName(entry.getKey()) + "(" + entry.getValue() + ")")
+                            .map(entry -> RoleNameMapper.getKoreanRoleName(entry.getKey()) + "(" + entry.getValue() + ")")
                             .collect(Collectors.joining(", "));
 
                     Optional<Role> userPotRole = potMemberRepository.findRoleByUserId(pot.getPotId(), userId);
@@ -693,104 +202,7 @@ public class MyPotServiceImpl implements MyPotService {
                 .collect(Collectors.toList());
     }
 
-    @Override
-    public MyPotTaskStatusResponseDto updateTaskStatus(Long potId, Long taskId, TaskboardStatus status) {
-        Taskboard taskboard = taskboardRepository.findById(taskId)
-                .orElseThrow(() -> new IllegalArgumentException("Taskboard not found with id: " + taskId));
 
-         //Taskboard가 해당 Pot에 속해 있는지 확인
-        if (!taskboard.getPot().getPotId().equals(potId)) {
-            throw new IllegalArgumentException("The taskboard does not belong to the specified pot.");
-        }
-
-        // 입력받은 status 값으로 업데이트
-        taskboard.setStatus(status);
-
-        // 변경 사항 저장
-        taskboardRepository.save(taskboard);
-
-        return taskboardConverter.toTaskStatusDto(taskboard, status);
-    }
-
-    @Override
-    public List<MyPotTaskPreViewResponseDto> getTasksFromDate(Long potId, LocalDate date) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-        // 해당 pot의 존재 여부 확인 및 사용자가 pot의 멤버인지 확인
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 팟입니다."));
-
-//        List<Taskboard> taskboards = taskboardRepository.findByPotAndDeadLineGreaterThanEqualOrderByDeadLineAsc(pot, date);
-        List<Taskboard> taskboards = taskboardRepository.findByPotAndDeadLine(pot,date);
-
-        // DTO로 변환
-        return taskboards.stream()
-                .map(taskboard -> {
-                    List<Task> tasks = taskRepository.findByTaskboard(taskboard); // Task 조회
-                    List<PotMember> participants = tasks.stream()
-                            .map(Task::getPotMember) // Task에서 PotMember 추출
-                            .distinct()
-                            .collect(Collectors.toList());
-
-                    return taskboardConverter.toDto(taskboard, participants);
-                })
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public List<MonthlyTaskDto> getMonthlyTasks(Long potId, int year, int month) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String email = authentication.getName();
-
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-        Pot pot = potRepository.findById(potId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 팟입니다."));
-
-
-        // 현재 사용자의 PotMember 정보 가져오기
-        PotMember currentPotMember = potMemberRepository.findByPotAndUser(pot, user)
-                .orElseThrow(() -> new IllegalArgumentException("해당 팟의 멤버가 아닙니다."));
-
-        // 해당 월의 시작일과 마지막 일 계산
-        LocalDate startDate = LocalDate.of(year, month, 1);
-        LocalDate endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
-
-        // 해당 월의 모든 Taskboard 조회
-        List<Taskboard> taskboards = taskboardRepository
-                .findByPotAndDeadLineBetweenOrderByDeadLineAsc(pot, startDate, endDate);
-
-        return taskboards.stream()
-                .map(taskboard -> {
-                    List<Task> tasks = taskRepository.findByTaskboard(taskboard);
-                    // 현재 사용자의 참여 여부 확인
-                    boolean isParticipating = tasks.stream()
-                            .map(Task::getPotMember)
-                            .anyMatch(potMember -> potMember.equals(currentPotMember));
-
-                    return MonthlyTaskDto.builder()
-                            .taskId(taskboard.getTaskboardId())
-                            .deadLine(taskboard.getDeadLine())
-                            .isParticipating(isParticipating)
-                            .build();
-                })
-                .collect(Collectors.toList());
-    }
-
-    private String getKoreanRoleName(String role) {
-        Map<String, String> roleToKoreanMap = Map.of(
-                "BACKEND", "백엔드",
-                "FRONTEND", "프론트엔드",
-                "DESIGN", "디자인",
-                "PLANNING", "기획"
-        );
-        return roleToKoreanMap.getOrDefault(role, "알 수 없음");
-    }
 
     public boolean isOwner(Long potId) {
         // 현재 로그인한 사용자 가져오기
