+++ src/main/java/stackpot/stackpot/pot/service/PotServiceImpl.java
@@ -36,6 +36,7 @@ import java.time.LocalDate;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 
@@ -689,6 +690,5 @@ public class PotServiceImpl implements PotService {
 
         return potConverter.toDto(pot, recruitmentDetails);
     }
-
 
 }
