+++ src/main/java/stackpot/stackpot/task/repository/TaskboardRepository.java
@@ -26,8 +26,6 @@ public interface TaskboardRepository extends JpaRepository<Taskboard, Long> {
         @Query("DELETE FROM Taskboard f WHERE f.pot.potId = :potId")
         void deleteByPotId(@Param("potId") Long potId);
 
-        List<Taskboard> findByPotAndDeadLineGreaterThanEqualOrderByDeadLineAsc(Pot pot, LocalDate date);
-
         @Query("SELECT t FROM Taskboard t WHERE t.pot = :pot AND t.deadLine BETWEEN :startDate AND :endDate ORDER BY t.deadLine ASC")
         List<Taskboard> findByPotAndDeadLineBetweenOrderByDeadLineAsc(
                 @Param("pot") Pot pot,
