+++ src/main/java/stackpot/stackpot/task/service/TaskService.java
@@ -1,0 +1,19 @@
+package stackpot.stackpot.task.service;
+
+import stackpot.stackpot.task.dto.*;
+import stackpot.stackpot.task.entity.enums.TaskboardStatus;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Map;
+
+public interface TaskService {
+    MyPotTaskResponseDto creatTask(Long potId, MyPotTaskRequestDto.create request);
+    Map<TaskboardStatus, List<MyPotTaskPreViewResponseDto>> preViewTask(Long potId);
+    MyPotTaskResponseDto viewDetailTask(Long potId, Long taskboardId);
+    void deleteTaskboard(Long potId, Long taskboardId);
+    MyPotTaskStatusResponseDto updateTaskStatus(Long potId, Long taskId, TaskboardStatus status);
+    List<MyPotTaskPreViewResponseDto> getTasksFromDate(Long potId, LocalDate date);
+    List<MonthlyTaskDto> getMonthlyTasks(Long potId, int year, int month);
+    MyPotTaskResponseDto modifyTask(Long potId, Long taskboardId, MyPotTaskRequestDto.create request);
+}
